+++ src/app/js/upload/upload.js
@@ -91,38 +91,10 @@
         }],
         selectedOwner: undefined,
 
-        visibilityOptions: (function() {
-          function tt(key, description) {
-            return '<div class="left">' +
-              '<span class="bold">' +key+ '</span>: ' +
-              description +
-              '</div>'
-          }
-          return [{
-            value:   'owner',
-            tooltip: tt('owner', 'Visible only to the owner (submitting user or members of indicated organization)' +
-              ' or anyone with the URI.')
-            }, {
-              value:   'public',
-              tooltip: tt('public', 'Visible to any visitor or client application.')
-            }];
-        })(),
-
+        visibilityOptions:  utl.visibilityOptions,
         selectedVisibility: undefined,
 
-        statusOptions: (function() {
-          var opts = [];
-          _.each(cfg.ontologyStatuses, function(description, key) {
-            opts.push({
-              value:   key,
-              tooltip: '<div class="left">' +
-              '<span class="bold">' +key+ '</span>: ' +
-              description +
-              '</div>'
-            })
-          });
-          return opts;
-        })(),
+        statusOptions:  utl.statusOptions,
         selectedStatus: undefined
       };