+++ ChangeLog.md
@@ -2,6 +2,9 @@
 
 * 2017-02-14:  3.2.1
   - fix #57 "m2r: check and avoid triple duplications"
+    Also, avoid duplicates when loading mapping triples in m2r interface in case registered
+    content contains such duplicates (which may be the case due to initial implementations).
+    
   - fix #75 ""No such ontology" for external case in mapped ontologies"
   
 * 2017-02-09:  3.2.0